unify path normalization utility (#6865)
