+++ core/lib.rs
@@ -15,6 +15,7 @@ mod es_isolate;
 mod flags;
 mod module_specifier;
 mod modules;
+mod normalize_path;
 mod ops;
 pub mod plugin_api;
 mod resources;
@@ -43,6 +44,7 @@ pub use crate::modules::ModuleLoader;
 pub use crate::modules::ModuleSource;
 pub use crate::modules::ModuleSourceFuture;
 pub use crate::modules::RecursiveModuleLoad;
+pub use crate::normalize_path::normalize_path;
 pub use crate::ops::Buf;
 pub use crate::ops::Op;
 pub use crate::ops::OpAsyncFuture;
